+++ packages/admin/plugin.js
@@ -35,7 +35,7 @@ K.Plugin({
 		},
 
 		panelUser: {'itemUser_admin': {order:-5}},
-		panelPlace: {'itemPlace_admin': {order:-5}},
+		panelPlace: {'itemPlace_admin': {order:10}},
 		//panelPlaceEdit: {'itemPlace_admin': {order:-5}},
 		//panelPlace:'itemPlace_admin',