+++ packages/edit/i18n/en.js
@@ -8,10 +8,12 @@ i18n.map('en', {
 	btn_del: "Delete",
 	btn_ren: "Rename",
 	btn_reload: "Reload",
+	btn_delplace: "Delete this place",
 
 	label_editplaces: "Edits",
 	label_owner: "Creator",
 
+	edit_error_editing: "Change each field and close to return to public view",
 	edit_error_notOwner: "Not Owner",
 	edit_error_notUrl: "Url Not Valid",
 	edit_error_notMultigeom: "Multi Geometries Not Supported to editing"