+++ packages/edit/i18n/it.js
@@ -5,9 +5,11 @@ i18n.map('it', {
 	btn_editinsert: "Crea Nuovo Luogo Qui!",
 	btn_edit: "Modifica",
 	btn_save: "Salva",
-	btn_del: "Rimuovi",
+	btn_del: "Elimina",
 	btn_ren: "Rinomina",
 	btn_reload: "Ricarica",
+	btn_delplace: "Elimina questo luogo",
 
-	label_editplaces: "Modifiche"
+	label_editplaces: "Modifiche",
+	edit_error_editing: "Modifica ogni campo e chiudi per tornare alla vista pubblica",	
 });