+++ packages/edit/plugin.js
@@ -6,11 +6,10 @@ K.Plugin({
 		//create new placeholder used from other plugins
 		//
 		panelPlaceEdit: {
-			'panelPlace_ui_btn_checkin': {order: -10},
-			'panelPlace_ui_btns': {order: -10},
-			'panelPlaceEdit_edit_map': {order: -10},
-			'panelPlaceEdit_edit_del': {order: -10},			
+			'panelPlaceEdit_edit_status': {order: -10},
 			'panelPlaceEdit_edit_info': {order: -10},
+			'panelPlaceEdit_edit_map': {order: -10},
+			'panelPlaceEdit_edit_del': {order: -10}
 		},
 		panelProfile: {
 			'panelProfile_edit': {order: 10 }