+++ packages/share/i18n/en.js
@@ -1,7 +1,7 @@
 
 i18n.map('en', {
 	btn_share: "Send",
-	btn_copied: "Copied!",
+	btn_copied: "Url Copied!",
 	btn_copyloc: "Copy location",
 	btn_nav: "Navigate",
 	btn_navapp: "Default App",