+++ packages/share/i18n/fr.js
@@ -1,7 +1,7 @@
 
 i18n.map('fr', {
 	btn_share: "Envoyer",
-	btn_copied: "Copié!",
+	btn_copied: "Url Copié!",
 	btn_nav: "Naviguer avec",
 	btn_navapp: "App de Navigation",
 	btn_google: "Google",